% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rprintn}
\alias{rprintn}
\title{Build a character vector or list with number-based
string formatting}
\usage{
rprintn(x, ...)
}
\arguments{
\item{x}{The character vector or list to be transformed}

\item{...}{The arguments that specify the set of values to be
placed}
}
\description{
The \code{rprintn} function applies number-based formatter to
transform the given character vector to specific texts with
numbers replaced by a given set of values in correct
formats.
}
\examples{
\dontrun{

# Format a single-entry character vector with numbering mechanism
rprintf("Hello, {1}", "world")
rprintf("{1} ({2} years old)","Ken",24)
rprintf("He is {1} but has a height of {2:.2f}cm",18,190)
rprintf("{1}, {2:.1f}, {3:+.2f}, {2}, {1:.0f}",1.56,2.34,3.78)
rprintf("{2},{1}","x","y")
}
}

